+++ package.json
@@ -1,5 +1,5 @@
 {
-    "type": "module",
+    "type": "CommonJS",
     "scripts": {
         "build": "tsc -p src/ofgemCheck/"
     },
