+++ src/RHIScraper/main/RHI/getRHIDetails.ts
@@ -2,7 +2,7 @@ import { Page } from "puppeteer-core";
 import * as cheerio from "cheerio";
 const PostcodesIO = require("postcodesio-client");
 const postcodes = new PostcodesIO("https://api.postcodes.io");
-import { RHIsTable } from "../../../../SmartSuite/tables.js";
+import { RHIsTable } from "../../../SmartSuite/tables.js";
 export default async function getRHIDetails(
     accountID: string,
     page: Page,
