+++ src/RHIScraper/main/RHI/validateLogin.ts
